Created tracking function (threshold)

+++ pylon_example/Tracker_horst/Pylon_with_OpenCV/threshold.h
@@ -5,4 +5,4 @@
 #include <opencv2/video/video.hpp>
 
 
-cv::Mat GetThresholdedImage(cv::Mat img,  uint8_t low_h, uint8_t high_h, uint8_t low_s, uint8_t high_s, uint8_t low_v, uint8_t high_v);
+cv::Mat GetThresholdedImage(cv::Mat img, uint low_h, uint high_h, uint low_s, uint high_s, uint low_v, uint high_v);
